feat: Move CommonAppBar to the bottom
