+++ src/components/BasePage.tsx
@@ -44,8 +44,8 @@ const BasePage = ({ children, pageName, needsAuth = true, isLoading = false }: B
     }
     return (
         <>
-            <CommonAppBar pageName={pageName} handleLogout={handleLogout} />
-            <Container component='main' maxWidth='xs' sx={{ pb: 4 }}>
+            <CommonAppBar handleLogout={handleLogout} />
+            <Container component='main' maxWidth='xs' sx={{ pb: 14 }}>
                 <CssBaseline />
                 {children}
             </Container>
